design:복약상세 카드 수정
